move summary code to node
