+++ electron-app/package.json
@@ -50,7 +50,8 @@
     "electron-devtools-installer": "^3.2.0",
     "electron-is-dev": "^2.0.0",
     "electron-squirrel-startup": "^1.0.0",
-    "fs-plus": "^3.1.1"
+    "fs-plus": "^3.1.1",
+    "arquero": "^5.1.0"
   },
   "devDependencies": {
     "@electron-forge/cli": "^6.0.0-beta.64",
