+++ package.json
@@ -83,7 +83,6 @@
   },
   "dependencies": {
     "@babel/helper-get-function-arity": "^7.16.7",
-    "arquero": "^5.1.0",
     "file-loader": "^6.2.0",
     "pg": "^8.7.1",
     "uuid": "^8.3.2",
